rename files to automations
