+++ src/graphql/Automations/agentConfig.gql
