+++ src/graphql/Automations/flowConfigs.gql
