+++ src/graphql/Automations/hooks.gql
