+++ src/pages/Dashboard/Automations/Automations.vue
@@ -1,12 +1,12 @@
 <script>
-import ActionCard from '@/pages/Dashboard/Actions/ActionCard'
-import AddActionCard from '@/pages/Dashboard/Actions/AddActionCard'
+import AutoCard from '@/pages/Dashboard/Automations/AutoCard'
+import AddAutoCard from '@/pages/Dashboard/Automations/AddAutoCard'
 import { mapGetters } from 'vuex'
 
 export default {
   components: {
-    ActionCard,
-    AddActionCard
+    AutoCard,
+    AddAutoCard
   },
   data() {
     return {
@@ -59,7 +59,7 @@ export default {
   apollo: {
     hooks: {
       query() {
-        return require('@/graphql/Actions/hooks.gql')
+        return require('@/graphql/Automations/hooks.gql')
       },
       variables() {
         return {}
@@ -76,9 +76,9 @@ export default {
   <div v-if="noAccess">
     <v-row>
       <v-col cols="12" class="text-center text-h6 pa-12">
-        You do not have access to Actions. If you'd like to be able to create
-        actions that notify you (or even cancel your run) when there's a change
-        in your agent or a run,
+        You do not have access to Automations. If you'd like to be able to
+        create automations that notify you (or even cancel your run) when
+        there's a change in your agent or a run,
         <router-link :to="{ name: 'plans' }">upgrade</router-link> your account.
       </v-col>
     </v-row>
@@ -86,12 +86,12 @@ export default {
   <div v-else>
     <v-row>
       <v-col>
-        <AddActionCard v-if="canEdit && !closeCard" @refresh="handleRefresh" />
+        <AddAutoCard v-if="canEdit && !closeCard" @refresh="handleRefresh" />
       </v-col>
     </v-row>
     <v-row>
       <v-col v-for="(hook, i) in sortedHooks" :key="i" cols="12">
-        <ActionCard
+        <AutoCard
           :hook="hook"
           :can-edit="canEdit"
           @open-edit="handleEdit(hook, i)"
